mpm/synapsis_lib/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synapsis"
version = "1.0.0"
description = "Gerador de mapas mentais com LLM injetável"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MPM Team"}
]
keywords = ["mindmap", "llm", "yaml", "visualization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
groq = ["groq>=0.4"]
openai = ["openai>=1.0"]

[project.urls]
Homepage = "https://github.com/mpm/synapsis"
Documentation = "https://github.com/mpm/synapsis#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["synapsis*"]

[tool.setuptools.package-data]
synapsis = ["../templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
